POC: starts to implement extensible parser
This parser allows the user to configure what "commands" are known and
supported by way of providing a list of "modules". The idea is that
users can then customize Crossplane to parse NGINX configurations with
specific module sets including their own that we know nothing about.

We could provide presets that contain specific modules/versions that can
be used with specific builds/versions of NGINX.

This POC has support for a very limited set of directives and a basic
capability to handle includes. Includes are currently handled similar to
what Crossplane does today, but I think I'm seeing issues with that
strategy. Specifically, the include's context is set when the file name
is first discovered. I think if that file is included a second time in a
different context, the context would be incorrect.