POC: Adds ability to ignore commands to parser
A list of commands can be passed to the parser that are then used to
match commands (directives) that the parser should "ignore" and not
include in the result. If an ignored command is a "block type" then the
parser recursively parses the block and ignores any error.

This is inline with existing Crossplane behavior but as others have
noted, it does seem a little weird as it changes the meaning/intent of
the NGINX configuration in a way that may no longer be valid. Arguably
it is up to the user to handle this use case some how.

I gave some thought to changing the parser "type" to be a pull model
similar to `Scanner` where the user could ask for the next directive and
on receiving an `include` could then choose to ignore it or even handle
it as the "single file" use case where the contents are injected into
the `Block` field of the parent directive instead of added to the file
list for later parsing. I'm not sure how that would look yet.