iPOC: Implements a partial Lua parser for content_by_lua_block
Added the concept of an "extension" to the Scanner that can be used to
implement alternative grammars such as Lua. Unlike the Python version of
crossplane, extensions to not need to be "registered" ahead of time.
Instead, when a token is detected for which you want to start scanning
with a different grammar, you then use the ScanWith method to continue
tokenizing the config.

This is using the same `Directive` type that already exists in
crossplane, but it starts to show that the existing struct may be
insufficient. Lua doesn't fit well into `Directive` so like the python
implementation of crossplane, the Lua script is jammed into the Args
field.